requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.23.8

# Machine learning and embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
faiss-cpu==1.7.4
huggingface-hub==0.17.3

# Image processing
Pillow==10.1.0

# LangChain and LLM
langchain==0.0.335
langchain-community==0.0.6

# Web interface
streamlit==1.28.1

# Utilities
numpy==1.24.3
requests==2.31.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1